requirements.txt
flask
flask-cors
numpy
python-dotenv
